Override default QC 2.4 Arbitrary instance
In previous QC versions we had to declare our own instances of Arbitrary for
Word8/Word16/Word32/Word64. This suited us well, as we want to implement our
own shrinking method. In this patch we're working around these instances with
the W datatype, a wrapper for the datatypes we're interested in.