Simplify internal of monad, implement Applicative.
Turned the monad into a simple state monad.
Implemented a data type 'Block' which represents a block of data that is
bounded in size and thus can be read with only one bound check.
Implemented Applicative for the 'Block' data type. GHC is even clever enough to
calculate the total size at compile time. Neat!
Read Blocks with the new 'block' function.