Support bit ordering, bit skipping and byte alignment
Bit ordering: this package now supports writing bits in different
orders. You can indicate if bytes are filled from MSB to LSB (the mode
we had until now) or vice-versa, and you can indicate if words you write
are stored with the MSB first (what we had until now) or with the LSB
first.

Bit skipping: you can now skip a given amount of bits in BitGet and
BitPut. In the latter, bits written are considered undefined (currently
we write zeroes but you shouldn't rely on this).

Byte alignment: you can now skip an undetermined number of bits in
BitGet and BitPut to align the cursor on the next byte. If the cursor is
already on a byte boundary, the cursor doesn't move.

MSB: most-significant bit(s)
LSB: least-significant bit(s)